app/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyMySQL==1.1.1
tzlocal==5.2
Werkzeug==3.1.3
zipp==3.21.0
